Centralize HAS_LIGER_KERNEL and shared Liger DPO helpers in llm_utils
